Initial commit of .bashrc
This is an initial commit of my personal .bashrc so
that it is more accessible.  The .bashrc is specifically
the stock Ubuntu 16.04 LTS .bashrc along with a few
useful aliases for ODL development.  Namely,

> odl-clone <project>

Easier cloning of odl projects.  For example,

> odl-clone aaa

Also, some nice ones for maven:

> mci

> mcis
Do a build but skip tests.

> mciq
Do a build but use the -Q profile.  Useful to quickly
produce artifacts.

Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>